MappingEditorRequestedEvents moved to conceptandrelationview package
